+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/vungle/mediation/VungleInterstitialAdapter.java
@@ -29,6 +29,7 @@ import androidx.annotation.Keep;
 import com.google.ads.mediation.vungle.VungleInitializer;
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.AdSize;
+import com.google.android.gms.ads.MediationUtils;
 import com.google.android.gms.ads.mediation.MediationAdRequest;
 import com.google.android.gms.ads.mediation.MediationBannerAdapter;
 import com.google.android.gms.ads.mediation.MediationBannerListener;
@@ -246,11 +247,18 @@ public class VungleInterstitialAdapter implements MediationInterstitialAdapter,
       return;
     }
 
-    adLayout = new RelativeLayout(context);
-    // Make adLayout wrapper match the requested ad size, as Vungle's ad uses MATCH_PARENT for
+    // Create the adLayout wrapper with the requested ad size, as Vungle's ad uses MATCH_PARENT for
     // its dimensions.
-    RelativeLayout.LayoutParams adViewLayoutParams = new RelativeLayout.LayoutParams(
-        adSize.getWidthInPixels(context), adSize.getHeightInPixels(context));
+    adLayout = new RelativeLayout(context);
+    int adLayoutHeight = adSize.getHeightInPixels(context);
+    // If the height is 0 (e.g. for inline adaptive banner requests), use the closest supported size
+    // as the height of the adLayout wrapper.
+    if (adLayoutHeight <= 0) {
+      float density = context.getResources().getDisplayMetrics().density;
+      adLayoutHeight = Math.round(adConfig.getAdSize().getHeight() * density);
+    }
+    RelativeLayout.LayoutParams adViewLayoutParams =
+        new RelativeLayout.LayoutParams(adSize.getWidthInPixels(context), adLayoutHeight);
     adLayout.setLayoutParams(adViewLayoutParams);
 
     mBannerRequest = mVungleManager
@@ -264,6 +272,8 @@ public class VungleInterstitialAdapter implements MediationInterstitialAdapter,
 
     mBannerRequest.setAdLayout(adLayout);
     mBannerRequest.setVungleListener(mVungleBannerListener);
+
+    Log.d(TAG, "Requesting banner with ad size: " + adConfig.getAdSize());
     mBannerRequest.requestBannerAd(context, config.getAppId());
   }
 
@@ -317,84 +327,35 @@ public class VungleInterstitialAdapter implements MediationInterstitialAdapter,
   }
 
   private boolean hasBannerSizeAd(Context context, AdSize adSize, AdConfig adConfig) {
-    ArrayList<AdConfig.AdSize> potentials = new ArrayList<>(4);
-    potentials.add(0, BANNER_SHORT);
-    potentials.add(1, BANNER);
-    potentials.add(2, BANNER_LEADERBOARD);
-    potentials.add(3, VUNGLE_MREC);
+    ArrayList<AdSize> potentials = new ArrayList<>();
+    potentials.add(new AdSize(BANNER_SHORT.getWidth(), BANNER_SHORT.getHeight()));
+    potentials.add(new AdSize(BANNER.getWidth(), BANNER.getHeight()));
+    potentials.add(new AdSize(BANNER_LEADERBOARD.getWidth(), BANNER_LEADERBOARD.getHeight()));
+    potentials.add(new AdSize(VUNGLE_MREC.getWidth(), VUNGLE_MREC.getHeight()));
 
-    Log.i(TAG, "Potential ad sizes: " + potentials.toString());
-    AdConfig.AdSize closestSize = findClosestSize(context, adSize, potentials);
+    AdSize closestSize = MediationUtils.findClosestSize(context, adSize, potentials);
     if (closestSize == null) {
       Log.i(TAG, "Not found closest ad size: " + adSize);
       return false;
     }
     Log.i(TAG,
-        "Found closest ad size: " + closestSize.toString() + " for request ad size:" + adSize);
+        "Found closest ad size: " + closestSize.toString() + " for requested ad size: " + adSize);
 
-    adConfig.setAdSize(closestSize);
+    if (closestSize.getWidth() == BANNER_SHORT.getWidth()
+        && closestSize.getHeight() == BANNER_SHORT.getHeight()) {
+      adConfig.setAdSize(BANNER_SHORT);
+    } else if (closestSize.getWidth() == BANNER.getWidth()
+        && closestSize.getHeight() == BANNER.getHeight()) {
+      adConfig.setAdSize(BANNER);
+    } else if (closestSize.getWidth() == BANNER_LEADERBOARD.getWidth()
+        && closestSize.getHeight() == BANNER_LEADERBOARD.getHeight()) {
+      adConfig.setAdSize(BANNER_LEADERBOARD);
+    } else if (closestSize.getWidth() == VUNGLE_MREC.getWidth()
+        && closestSize.getHeight() == VUNGLE_MREC.getHeight()) {
+      adConfig.setAdSize(VUNGLE_MREC);
+    }
 
     return true;
   }
-
-  // Copied some code from FB adapter:
-  // https://github.com/googleads/googleads-mobile-android-mediation/blob/ebce3b3ccf1c7a0cd8ecb31819c0037b8885d584/
-  // ThirdPartyAdapters/facebook/facebook/src/main/java/com/google/ads/mediation/facebook/FacebookAdapter.java#L760
-
-  // Start of helper code to remove when available in SDK
-
-  /**
-   * Find the closest supported AdSize from the list of potentials to the provided size. Returns
-   * null if none are within given threshold size range.
-   */
-  private AdConfig.AdSize findClosestSize(
-      Context context, AdSize original, ArrayList<AdConfig.AdSize> potentials) {
-    if (potentials == null || original == null) {
-      return null;
-    }
-    float density = context.getResources().getDisplayMetrics().density;
-    int actualWidth = Math.round(original.getWidthInPixels(context) / density);
-    int actualHeight = Math.round(original.getHeightInPixels(context) / density);
-    original = new AdSize(actualWidth, actualHeight);
-
-    AdConfig.AdSize largestPotential = null;
-    for (AdConfig.AdSize potential : potentials) {
-      if (isSizeInRange(original, potential)) {
-        if (largestPotential == null) {
-          largestPotential = potential;
-        } else {
-          largestPotential = getLargerByArea(largestPotential, potential);
-        }
-      }
-    }
-    return largestPotential;
-  }
-
-  private static boolean isSizeInRange(AdSize original, AdConfig.AdSize potential) {
-    if (potential == null) {
-      return false;
-    }
-    double minWidthRatio = 0.5;
-    double minHeightRatio = 0.7;
-
-    int originalWidth = original.getWidth();
-    int potentialWidth = potential.getWidth();
-    int originalHeight = original.getHeight();
-    int potentialHeight = potential.getHeight();
-
-    if (originalWidth * minWidthRatio > potentialWidth || originalWidth < potentialWidth) {
-      return false;
-    }
-
-    return !(originalHeight * minHeightRatio > potentialHeight)
-        && originalHeight >= potentialHeight;
-  }
-
-  private static AdConfig.AdSize getLargerByArea(AdConfig.AdSize size1, AdConfig.AdSize size2) {
-    int area1 = size1.getWidth() * size1.getHeight();
-    int area2 = size2.getWidth() * size2.getHeight();
-    return area1 > area2 ? size1 : size2;
-  }
-  // End code to remove when available in SDK
 
 }