+++ ThirdPartyAdapters/verizonmedia/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Verizon Media Android Mediation Adapter Changelog
 
+#### Next Version
+- Updated the adapter to support inline adaptive banner requests.
+
 #### Version 1.6.0.0
 - Support for Verizon Media SDK v1.6.0.
 - Updated the minimum required Google Mobile Ads SDK version to 19.1.0.